+++ plone/app/standardtiles/existingcontent.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_parent
-from Products.CMFPlone.utils import safe_unicode
-from ZODB.POSException import POSKeyError
+from operator import itemgetter
 from plone import api
 from plone.api.exc import InvalidParameterError
 from plone.app.blocks import utils
@@ -11,16 +10,23 @@ from plone.app.standardtiles import PloneMessageFactory as _
 from plone.app.z3cform.widget import RelatedItemsFieldWidget
 from plone.autoform import directives as form
 from plone.memoize.view import memoize
+from plone.registry.interfaces import IRegistry
 from plone.supermodel import model
 from plone.tiles import Tile
 from plone.uuid.interfaces import IUUID
+from Products.CMFPlone.utils import safe_unicode
 from repoze.xmliter.utils import getHTMLSerializer
 from z3c.form import validator
 from zExceptions import Unauthorized
+from ZODB.POSException import POSKeyError
 from zope import schema
 from zope.browser.interfaces import IBrowserView
+from zope.component import getUtility
 from zope.component.hooks import getSite
 from zope.interface import Invalid
+from zope.interface import provider
+from zope.schema.interfaces import IVocabularyFactory
+from zope.schema.vocabulary import SimpleVocabulary
 
 import six
 
@@ -69,25 +75,16 @@ class IExistingContentTile(model.Schema):
         'content_uid',
         RelatedItemsFieldWidget,
         vocabulary='plone.app.vocabularies.Catalog',
-        pattern_options={
-            'recentlyUsed': True,
-        }
+        pattern_options={'recentlyUsed': True},
     )
 
-    show_title = schema.Bool(
-        title=_(u'Show content title'),
-        default=True,
-    )
+    show_title = schema.Bool(title=_(u'Show content title'), default=True)
 
     show_description = schema.Bool(
-        title=_(u'Show content description'),
-        default=True
+        title=_(u'Show content description'), default=True
     )
 
-    show_text = schema.Bool(
-        title=_(u'Show content text'),
-        default=True,
-    )
+    show_text = schema.Bool(title=_(u'Show content text'), default=True)
 
     show_image = schema.Bool(
         title=_(u'Show content image (if available)'),
@@ -111,9 +108,16 @@ class IExistingContentTile(model.Schema):
         title=_(u'Tile additional styles'),
         description=_(
             u'Insert a list of additional CSS classes that will',
-            ' be added to the tile'),
+            ' be added to the tile',
+        ),
         default=u'',
         required=False,
+    )
+
+    view_template = schema.Choice(
+        title=_(u'Display mode'),
+        source=_(u'Available Content Views'),
+        required=True,
     )
 
 
@@ -130,8 +134,7 @@ class SameContentValidator(validator.SimpleFieldValidator):
 
 # Register validator
 validator.WidgetValidatorDiscriminators(
-    SameContentValidator,
-    field=IExistingContentTile['content_uid'],
+    SameContentValidator, field=IExistingContentTile['content_uid']
 )
 
 
@@ -159,34 +162,36 @@ class ExistingContentTile(Tile):
 
     @property
     @memoize
-    def default_view(self):
+    def content_view(self):
         context = self.content_context
         if context is not None:
-            item_layout = context.getLayout()
-            default_view = context.restrictedTraverse(item_layout)
-            return default_view
+            view_name = self.data.get('view_template') or context.getLayout()
+            return api.content.get_view(
+                name=view_name, context=context, request=self.request
+            )
         return None
 
     @property
     def item_macros(self):
-        default_view = self.default_view
-        if default_view and IBrowserView.providedBy(default_view):
+        view = self.content_view
+        if view and IBrowserView.providedBy(view):
             # IBrowserView
-            if getattr(default_view, 'index', None):
-                return default_view.index.macros
-        elif default_view:
+            if getattr(view, 'index', None):
+                return view.index.macros
+        elif view:
             # FSPageTemplate
-            return default_view.macros
+            return view.macros
         return None
 
     @property
     def item_panels(self):
-        default_view = self.default_view
-        html = default_view()
+        content_view = self.content_view
+        html = content_view()
         if isinstance(html, six.text_type):
             html = html.encode('utf-8')
-        serializer = getHTMLSerializer([html], pretty_print=False,
-                                       encoding='utf-8')
+        serializer = getHTMLSerializer(
+            [html], pretty_print=False, encoding='utf-8'
+        )
         panels = dict(
             (node.attrib['data-panel'], node)
             for node in utils.panelXPath(serializer.tree)
@@ -199,9 +204,15 @@ class ExistingContentTile(Tile):
             except RuntimeError:  # maximum recursion depth exceeded
                 return []
             clear = '<div style="clear: both;"></div>'
-            return [''.join([safe_unicode(serializer.serializer(child))
-                             for child in node.getchildren()])
-                    for name, node in panels.items()] + [clear]
+            return [
+                ''.join(
+                    [
+                        safe_unicode(serializer.serializer(child))
+                        for child in node.getchildren()
+                    ]
+                )
+                for name, node in panels.items()
+            ] + [clear]
         return []
 
     @property
@@ -211,9 +222,7 @@ class ExistingContentTile(Tile):
             return ''
         try:
             scale_view = api.content.get_view(
-                name='images',
-                context=context,
-                request=self.request,
+                name='images', context=context, request=self.request
             )
             scale = self.data.get('image_scale', 'thumb')
             return scale_view.scale('image', scale=scale).tag()
@@ -241,13 +250,28 @@ class ExistingContentTile(Tile):
     def __getattr__(self, name):
         # proxy attributes for this view to the selected view of the content
         # item so views work
-        if name in ('data',
-                    'content_context',
-                    'default_view',
-                    'item_macros',
-                    'item_panels',
-                    'getPhysicalPath',
-                    'index_html',
-                    ) or name.startswith(('_', 'im_', 'func_')):
+        if name in (
+            'data',
+            'content_context',
+            'content_view',
+            'item_macros',
+            'item_panels',
+            'getPhysicalPath',
+            'index_html',
+        ) or name.startswith(('_', 'im_', 'func_')):
             return Tile.__getattr__(self, name)
-        return getattr(self.default_view, name)
+        return getattr(self.content_view, name)
+
+
+@provider(IVocabularyFactory)
+def availableContentViewsVocabulary(context):
+    """Get available views for a content as vocabulary"""
+
+    registry = getUtility(IRegistry)
+    listing_views = (
+        registry.get('plone.app.standardtiles.content_views', {}) or {}
+    )
+    voc = [SimpleVocabulary.createTerm('', '', 'Default view')]
+    for key, label in sorted(listing_views.items(), key=itemgetter(1)):
+        voc.append(SimpleVocabulary.createTerm(key, key, label))
+    return SimpleVocabulary(voc)