+++ plone/app/standardtiles/tests/test_existing_content.py
@@ -24,12 +24,18 @@ import transaction
 
 
 def image():
-    img = Image.new('RGB', (random.randint(320, 640),
-                            random.randint(320, 640)))
+    img = Image.new(
+        'RGB', (random.randint(320, 640), random.randint(320, 640))
+    )
     draw = ImageDraw.Draw(img)
-    draw.rectangle(((0, 0), img.size), fill=(random.randint(0, 255),
-                                             random.randint(0, 255),
-                                             random.randint(0, 255)))
+    draw.rectangle(
+        ((0, 0), img.size),
+        fill=(
+            random.randint(0, 255),
+            random.randint(0, 255),
+            random.randint(0, 255),
+        ),
+    )
     del draw
     output = six.BytesIO()
     img.save(output, 'PNG')
@@ -48,15 +54,17 @@ class ExistingContentTileTests(TestCase):
         self.browser.handleErrors = False
         self.browser.addHeader(
             'Authorization',
-            'Basic %s:%s' % (TEST_USER_NAME, TEST_USER_PASSWORD,)
+            'Basic %s:%s' % (TEST_USER_NAME, TEST_USER_PASSWORD),
         )
 
         self.unprivileged_browser = Browser(self.layer['app'])
 
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
         page_id = self.portal.invokeFactory(
-            'Document', 'a-simple-page',
-            title=u'A simple page', description=u'A description'
+            'Document',
+            'a-simple-page',
+            title=u'A simple page',
+            description=u'A description',
         )
         self.page = self.portal[page_id]
         self.pageURL = self.portal[page_id].absolute_url()
@@ -69,9 +77,11 @@ class ExistingContentTileTests(TestCase):
 
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
-            title=u'An another page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'an-another-page',
+            title=u'An another page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         self.portal[page_id].text = RichTextValue(u'Hello World!')
 
@@ -80,10 +90,10 @@ class ExistingContentTileTests(TestCase):
         transaction.commit()
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_text=True',
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique'
+            '?content_uid={page_uuid}&show_text=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertIn(u'Hello World!', self.unprivileged_browser.contents)
@@ -92,9 +102,11 @@ class ExistingContentTileTests(TestCase):
         """
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
-            title=u'An another page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'an-another-page',
+            title=u'An another page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         self.portal[page_id].text = RichTextValue(u'Hello World!')
 
@@ -102,29 +114,34 @@ class ExistingContentTileTests(TestCase):
 
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_title=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_title=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertIn(u'An another page', self.unprivileged_browser.contents)
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
-        self.assertNotIn(u'An another page', self.unprivileged_browser.contents)
+        self.assertNotIn(
+            u'An another page', self.unprivileged_browser.contents
+        )
 
     def test_existing_content_tile_show_description(self):
         """
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
-            title=u'An another page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'an-another-page',
+            title=u'An another page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         self.portal[page_id].text = RichTextValue(u'Hello World!')
 
@@ -132,18 +149,19 @@ class ExistingContentTileTests(TestCase):
 
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_description=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_description=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertIn(u'A description', self.unprivileged_browser.contents)
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertNotIn(u'A description', self.unprivileged_browser.contents)
@@ -152,9 +170,11 @@ class ExistingContentTileTests(TestCase):
         """
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
-            title=u'An another page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'an-another-page',
+            title=u'An another page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         self.portal[page_id].text = RichTextValue(u'Hello World!')
 
@@ -162,18 +182,19 @@ class ExistingContentTileTests(TestCase):
 
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_text=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_text=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertIn(u'Hello World!', self.unprivileged_browser.contents)
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertNotIn(u'Hello World!', self.unprivileged_browser.contents)
@@ -182,32 +203,35 @@ class ExistingContentTileTests(TestCase):
         """
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'a-page',
-            title=u'A page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'a-page',
+            title=u'A page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         image_id = self.portal.invokeFactory(
-            'Image', 'an-image',
-            title=u'An Image', description=u'foo',
-            image=NamedImage(image(), 'image/png', filename=u'color.png')
+            'Image',
+            'an-image',
+            title=u'An Image',
+            description=u'foo',
+            image=NamedImage(image(), 'image/png', filename=u'color.png'),
         )
         page_uuid = IUUID(self.portal[page_id])
         image_uuid = IUUID(self.portal[image_id])
 
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_image=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_image=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
         self.assertNotIn(u'<img src="', self.unprivileged_browser.contents)
-
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            image_uuid +
-            '&show_image=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={image_uuid}&show_image=True'.format(
+                portal_url=self.portalURL, image_uuid=image_uuid
+            )
         )
 
         self.assertIn(u'<img src="', self.unprivileged_browser.contents)
@@ -221,18 +245,20 @@ class ExistingContentTileTests(TestCase):
         settings.globally_enabled = True
 
         page_id = self.portal.invokeFactory(
-            'Document', 'a-commented-page',
-            title=u'A commented page', description=u'A description',
-            text=u'Hello World!'
+            'Document',
+            'a-commented-page',
+            title=u'A commented page',
+            description=u'A description',
+            text=u'Hello World!',
         )
         page = self.portal[page_id]
         page_uuid = IUUID(page)
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_comments=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_comments=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
         self.assertIn(u'0', self.unprivileged_browser.contents)
 
@@ -248,10 +274,10 @@ class ExistingContentTileTests(TestCase):
         conversation.addComment(comment1)
         transaction.commit()
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_comments=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_comments=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
         self.assertIn(u'1', self.unprivileged_browser.contents)
 
@@ -260,10 +286,12 @@ class ExistingContentTileTests(TestCase):
         the content linked to existing content tile, the tile renders
         empty"""
         self.portal.portal_workflow.setDefaultChain(
-            'simple_publication_workflow')
+            'simple_publication_workflow'
+        )
 
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
+            'Document',
+            'an-another-page',
             title=u'An another page',
             description=u'A description',
             text=RichTextValue(u'Hello World!'),
@@ -274,9 +302,10 @@ class ExistingContentTileTests(TestCase):
 
         self.unprivileged_browser.handleErrors = False
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertNotIn(u'Hello World!', self.unprivileged_browser.contents)
@@ -296,7 +325,7 @@ class ExistingContentTileTests(TestCase):
         transaction.commit()
 
         self.browser.open(
-            '{}/@@edit-tile/plone.app.standardtiles.existingcontent/unique'.format(
+            '{}/@@edit-tile/plone.app.standardtiles.existingcontent/unique'.format(  # noqa
                 page.absolute_url()
             )
         )
@@ -308,9 +337,12 @@ class ExistingContentTileTests(TestCase):
         self.assertIn(u'not select the same content', self.browser.contents)
 
         page2_id = self.portal.invokeFactory(
-            'Document', 'an-another-page-2',
-            title=u'An another page', description=u'A description',
-            text=u'Hello World!')
+            'Document',
+            'an-another-page-2',
+            title=u'An another page',
+            description=u'A description',
+            text=u'Hello World!',
+        )
         page2 = self.portal[page2_id]
         page2_uuid = IUUID(page2)
         page2.text = RichTextValue(u'Hello World!')
@@ -330,8 +362,10 @@ class ExistingContentTileTests(TestCase):
 
         """
         page_id = self.portal.invokeFactory(
-            'Document', 'an-another-page',
-            title=u'An another page', description=u'A description'
+            'Document',
+            'an-another-page',
+            title=u'An another page',
+            description=u'A description',
         )
 
         page_uuid = IUUID(self.portal[page_id])
@@ -339,20 +373,60 @@ class ExistingContentTileTests(TestCase):
         transaction.commit()
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_title=True'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_title=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertNotIn(u'extra-class', self.unprivileged_browser.contents)
 
         self.unprivileged_browser.open(
-            self.portalURL +
-            '/@@plone.app.standardtiles.existingcontent/unique?content_uid=' +
-            page_uuid +
-            '&show_title=True' +
-            '&tile_class=extra-class'
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_title=True&tile_class=extra-class'.format(  # noqa
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
         )
 
         self.assertIn(u'extra-class', self.unprivileged_browser.contents)
+
+    def test_existing_content_tile_custom_layout(self):
+        """
+        Test that tile shows a custom layout, if set.
+        If not set, it uses the default content layout
+        """
+        page_id = self.portal.invokeFactory(
+            'Document',
+            'a-page-for-test',
+            title=u'An another page',
+            description=u'A description',
+        )
+        self.portal[page_id].text = RichTextValue(u'Hello World!')
+
+        page_uuid = IUUID(self.portal[page_id])
+
+        transaction.commit()
+
+        self.unprivileged_browser.open(
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_text=True&view_template='
+            'custom_existingcontent_layout'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
+        )
+        self.assertIn(
+            u'This is a custom layout', self.unprivileged_browser.contents
+        )
+        self.assertIn(u'Hello World!', self.unprivileged_browser.contents)
+
+        self.unprivileged_browser.open(
+            '{portal_url}/@@plone.app.standardtiles.existingcontent/unique?'
+            'content_uid={page_uuid}&show_text=True'.format(
+                portal_url=self.portalURL, page_uuid=page_uuid
+            )
+        )
+
+        self.assertNotIn(
+            u'This is a custom layout', self.unprivileged_browser.contents
+        )
+        self.assertIn(u'Hello World!', self.unprivileged_browser.contents)